+++ npbench/benchmarks/polybench/deriche/deriche.py
@@ -3,7 +3,7 @@
 import numpy as np
 
 
-def initialize(W, H, datatype=np.float64):
+def initialize(W, H, datatype=np.float32):
     alpha = datatype(0.25)
     imgIn = np.fromfunction(lambda i, j:
                             ((313 * i + 991 * j) % 65536) / 65535.0, (W, H),
